Form config abstraction
